requirements.txt
python-dotenv
numpy
matplotlib
seaborn
scikit-learn
pymongo
flask
requests
pandas
certifi
pymongo[srv]
pyaml
dill
mlflow
## -e .